pyproject.toml
[project]
name = "credit-ocr-demo-backend"
version = "0.1.0"
description = "Backend for credit request OCR demo"
authors = [
    {name = "Markus Kuehnle", email = "markus.kuehnle@example.com"},
]
dependencies = [
    "azure-ai-formrecognizer>=3.3.3",
    "dotenv>=0.9.9",
    "pyhocon>=0.3.61",
    "pytest>=8.4.0",
    "testcontainers>=4.10.0",
    "pytest-asyncio>=0.23.5",
    "aiohttp>=3.9.3",
    "pytest-ordering>=0.6",
    "pdf2image>=1.17.0",
    "Pillow>=10.2.0",
    "reportlab>=4.4.1",
]
requires-python = ">=3.10"
readme = "README.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/credit_ocr_demo_backend"]

[tool.hatch.build]
sources = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
